Make MediaWiki core ignore extension submodules
Using the MW release branches includes many extensions as submodules.
This causes issues when you want to use different versions of those
extensions. Checking out a different version of an extension tracked as
a submodule causes doing "git status" on MW core to show changes. This
causes issues when trying to use Ansible Git module, which needs a Git
repo.

This change makes Git ignore changes to submodules.

This has been in the Meza 31.x branch for a while, but was not
previously needed in the 27.x branch. The upgrade from MediaWiki 1.27.4
to 1.27.5 drove the need for this.